+++ scripts/preinstall.js
@@ -1,1 +1,2 @@
 require('./register-local-deps');
+process.exit(0);
